Reshape the dynamo helpers
