+++ lib/dynamodb/actions/save.js
@@ -1,4 +1,4 @@
-import { normalizeProject } from './normalize';
+import { normalizeProject } from '../tables/projects';
 
 export const save = (dynamo, table, data, cb = () => {}) => {
   const params = {
