+++ lib/dynamodb/index.js
@@ -1,0 +1,1 @@
+export { save, saveProject } from './actions/save';
