+++ lib/dynamodb/tables/projects.js
@@ -1,47 +1,12 @@
 import uuid from 'uuid';
 
-const filterFloat = value => {
-  if (/^(-|\+)?([0-9]+(\.[0-9]+)?|Infinity)$/.test(value)) return Number(value);
-  return NaN;
-};
-
-export const normalizeString = source =>
-  source && typeof source === 'string' && source.trim().length > 0
-    ? {
-        S: source.trim(),
-      }
-    : null;
-
-export const normalizeNumber = source => ({
-  N: (filterFloat(source) || 0).toString(),
-});
-
-export const normalizeStringSet = source => {
-  if (Array.isArray(source) && source.length > 0) {
-    const normalizedSource = source
-      .map(item => item.trim())
-      .filter(item => item !== '');
-
-    if (normalizedSource.length > 0) {
-      return {
-        SS: normalizedSource,
-      };
-    }
-  }
-
-  return null;
-};
-
-export const normalizeMap = source => ({
-  M: source,
-});
-
-export const normalizeList = source =>
-  Array.isArray(source) && source.length > 0
-    ? {
-        L: source,
-      }
-    : null;
+import {
+  normalizeList,
+  normalizeMap,
+  normalizeNumber,
+  normalizeString,
+  normalizeStringSet,
+} from '../helpers/normalize';
 
 export const normalizeProject = record => {
   const normalizedObject = {
@@ -103,3 +68,5 @@ export const normalizeProject = record => {
 
   return normalizedObject;
 };
+
+export default normalizeProject;
